pyproject.toml
[project]
name = "mcp-secure-server"
version = "1.0.0"
description = "Production-ready, security-first MCP server with zero data leakage"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "httpx>=0.27",
    "cachetools>=6.2.2",
]

[dependency-groups]
dev = [
    "pyright~=1.1.407",
    "pytest~=9.0.2",
    "pytest-cov",
    "pytest-asyncio",
    "ruff",
]

[build-system]
requires = ["uv_build>=0.8.11"]
build-backend = "uv_build"

[project.scripts]
mcp-secure = "mcp_secure_server.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing --cov-fail-under=95"

# RUFF settings
[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = [
  # https://docs.astral.sh/ruff/rules/
    "RUF",  # Ruff specific rules - https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
    "C",    # mccabe rules - https://pypi.org/project/mccabe/
    "F",    # pyflakes rules - https://pypi.org/project/pyflakes/
    "E",    # pycodestyle error rules - https://pypi.org/project/pycodestyle/
    "W",    # pycodestyle warning rules - https://pypi.org/project/pycodestyle/
    "N",    # Naming - https://pypi.org/project/pep8-naming/
    "A",    # builtins - https://pypi.org/project/flake8-builtins/
    "B",    # flake8-bugbear rules - https://pypi.org/project/flake8-bugbear/
    "I",    # isort rules - https://pypi.org/project/isort/
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit (security)
    "LOG",  # logging rules - https://pypi.org/project/flake8-logging/
    "PT",   # PyTest - https://pypi.org/project/flake8-pytest-style/
    "INP",  # no-pep420 - https://pypi.org/project/flake8-no-pep420/
    "ANN",  # annotation rules - https://pypi.org/project/flake8-annotations/
    "UP",   # pyupgrade - https://pypi.org/project/pyupgrade/
    "RET",  # returns - https://pypi.org/project/flake8-return/
    "SIM",  # simplify - https://pypi.org/project/flake8-simplify/
    "PL",   # PyLint (ALL) - https://pypi.org/project/pylint/
    "PLC",  # PyLint Conventions - https://pypi.org/project/pylint/
    "PLE",  # PyLint Error - https://pypi.org/project/pylint/
    "PLR",  # PyLint Refactor - https://pypi.org/project/pylint/
    "PLW",  # PyLint Warnings - https://pypi.org/project/pylint/
    "FURB", # refurb - https://pypi.org/project/refurb/
    "ASYNC",# async - https://pypi.org/project/flake8-async/
]
ignore = [
    # "C901",  # max-complexity-10
    "E501",  # line-too-long
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "PLC0415",  # import-outside-top-level (PLC0415)
    "S101", "S105", "S106"  # Allow for assert statements and test password strings
]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true
ignore-fully-untyped = true
suppress-dummy-args = true

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# Rumdl settings
# This is a tool for formatting markdown files, it can be used to fix issues automatically.
[tool.rumdl]
line-length = 100
exclude = [".amazonq/**",]
